+++ eng/scripts/Language-Settings.ps1
@@ -142,7 +142,8 @@ function Publish-dotnet-GithubIODocs ($DocLocation, $PublicArtifactLocation)
   Upload-Blobs -DocDir "$($DocsStagingDir)" -PkgName $pkgProperties.PackageId -DocVersion $pkgProperties.PackageVersion -ReleaseTag $pkgProperties.ReleaseTag
 }
 
-function Get-dotnet-GithubIoDocIndex() {
+function Get-dotnet-GithubIoDocIndex()
+{
   # Update the main.js and docfx.json language content
   UpdateDocIndexFiles -appTitleLang ".NET"
   # Fetch out all package metadata from csv file.
@@ -157,7 +158,8 @@ function Get-dotnet-GithubIoDocIndex() {
 
 # details on CSV schema can be found here
 # https://review.docs.microsoft.com/en-us/help/onboard/admin/reference/dotnet/documenting-nuget?branch=master#set-up-the-ci-job
-function Update-dotnet-CIConfig($pkgs, $ciRepo, $locationInDocRepo, $monikerId=$null){
+function Update-dotnet-CIConfig($pkgs, $ciRepo, $locationInDocRepo, $monikerId=$null)
+{
   $csvLoc = (Join-Path -Path $ciRepo -ChildPath $locationInDocRepo)
   
   if (-not (Test-Path $csvLoc)) {
@@ -198,7 +200,6 @@ function Update-dotnet-CIConfig($pkgs, $ciRepo, $locationInDocRepo, $monikerId=$
   Set-Content -Path $csvLoc -Value $allCSVRows
 }
 
-
 # function is used to auto generate API View
 function Find-dotnet-Artifacts-For-Apireview($artifactDir, $packageName = "")
 {
@@ -211,4 +212,14 @@ function Find-dotnet-Artifacts-For-Apireview($artifactDir, $packageName = "")
   }
   $packages = @{ $pkg.Name = $pkg.FullName }
   return $packages
+}
+
+function SetPackageVersion ($PackageName, $Version, $ServiceDirectory, $ReleaseDate, $BuildType=$null, $GroupId=$null)
+{
+  if($null -eq $ReleaseDate)
+  {
+    $ReleaseDate = Get-Date -Format "yyyy-MM-dd"
+  }
+  & "$EngDir/scripts/Update-PkgVersion.ps1" -ServiceDirectory $ServiceDirectory -PackageName $PackageName `
+  -NewVersionString $Version -ReleaseDate $ReleaseDate
 }