+++ sdk/servicebus/Microsoft.Azure.WebJobs.Extensions.ServiceBus/src/Bindings/CollectorValueProvider.cs
@@ -1,0 +1,43 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT License. See License.txt in the project root for license information.
+
+using System;
+using System.Threading.Tasks;
+using Microsoft.Azure.WebJobs.Host.Bindings;
+
+namespace Microsoft.Azure.WebJobs.ServiceBus.Bindings
+{
+    internal class CollectorValueProvider : IValueProvider
+    {
+        private readonly ServiceBusEntity _entity;
+        private readonly object _value;
+        private readonly Type _valueType;
+
+        public CollectorValueProvider(ServiceBusEntity entity, object value, Type valueType)
+        {
+            if (value != null && !valueType.IsAssignableFrom(value.GetType()))
+            {
+                throw new InvalidOperationException("value is not of the correct type.");
+            }
+
+            _entity = entity;
+            _value = value;
+            _valueType = valueType;
+        }
+
+        public Type Type
+        {
+            get { return _valueType; }
+        }
+
+        public Task<object> GetValueAsync()
+        {
+            return Task.FromResult(_value);
+        }
+
+        public string ToInvokeString()
+        {
+            return _entity.MessageSender.Path;
+        }
+    }
+}